refactor : hubManager 리팩토링
- 불필요한 주석 삭제
- adapter 분리
- Hubmanager이벤트 로직과 서비스 로직 분리
