+++ hub-service/src/main/java/com/gaebal_easy/client/hub/application/dto/HubRouteDto.java
@@ -1,15 +1,10 @@
 package com.gaebal_easy.client.hub.application.dto;
 
-import com.gaebal_easy.client.hub.domain.entity.VisitHub;
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.OneToMany;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @NoArgsConstructor
