+++ hub-service/src/main/java/com/gaebal_easy/client/hub/application/service/EventErrorHandler.java
@@ -1,17 +1,17 @@
 package com.gaebal_easy.client.hub.application.service;
 
+import com.gaebal_easy.client.hub.presentation.adapter.out.HubManagerEventConsumer;
 import gaebal_easy.common.global.dto.BaseMessage;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
-@Component
+@Service
 @Slf4j
 @RequiredArgsConstructor
 public class EventErrorHandler {
 
-    private final KafkaTemplate<String, Object> kafkaTemplate;
+    private final HubManagerEventConsumer hubManagerEventConsumer;
 
     public <T extends BaseMessage> void handleEventError(
             Exception exception,
@@ -21,9 +21,7 @@ public class EventErrorHandler {
         if(message.getErrorLocation()==null){
             message.setErrorInfo("hub-service", exception.getMessage());
         }
-        log.error("이벤트 처리 중 오류 발생: {}", exception.getMessage(), exception);
-        kafkaTemplate.send(errorTopic, message);
-
+        hubManagerEventConsumer.sendErrorEvent(message, errorTopic);
     }
 
 }
