+++ hub-service/src/main/java/com/gaebal_easy/client/hub/application/service/HubService.java
@@ -6,7 +6,6 @@ import com.gaebal_easy.client.hub.application.dto.checkStockDto.CheckStockRespon
 import com.gaebal_easy.client.hub.application.dto.checkStockDto.CheckStokProductDto;
 import com.gaebal_easy.client.hub.application.dto.ProductResponseDto;
 import com.gaebal_easy.client.hub.application.dto.checkStockDto.CheckStockDto;
-import com.gaebal_easy.client.hub.application.dto.checkStockDto.CheckStokProductDto;
 import com.gaebal_easy.client.hub.application.dto.checkStockDto.ReservationDto;
 import com.gaebal_easy.client.hub.domain.entity.Hub;
 import com.gaebal_easy.client.hub.domain.entity.HubProductList;
